+++ quick_sort/quick_sort.h
@@ -6,8 +6,9 @@
 # include <vector>
 
 namespace Sort {
-   void quick (std::vector<int>& v, unsigned l, unsigned r);
-   unsigned partition (std::vector<int>& v, unsigned l, unsigned r);
+   void quick (std::vector<long>& v, unsigned l, unsigned r);
+   unsigned partition (std::vector<long>& v, unsigned l, unsigned r);
 }
 
 #endif
+
